Website/requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6

# Document Processing
pypdf==3.15.1
python-docx==0.8.11
docx2txt==0.8
pdfplumber==0.9.0
python-pptx==0.6.21

# Data Processing
pandas==2.0.0
openpyxl==3.1.2

# Optional dependencies 
pdftotext==2.2.2  # Requires poppler installation

# Aspose libraries (require licenses for production use)
aspose-words==23.6.0
aspose-slides==23.6.0 